requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Django==3.0.5
django-cors-headers==3.2.1
djangorestframework==3.11.0
idna==2.9
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.0
pycparser==2.20
pyOpenSSL==19.1.0
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
